Update README.md
Fixed wrong worst-case performance estimates.

In the worst case, the tree has to be completely rebuilt. But since it has n-1 branches for n leaves, the building of the branches is just a constant factor (no unnecessary branches are being created). So worst case runtime is O(n) instead of O(n log n)